GM seed and beta 7 changes
<em>Update for GMSeed: there are two big problems since I originally
posted this, both related to optionals. One is that optionals are not
considered boolean values so <code>if myOptional{do someting}</code>
must be <code>if (myOptional != nil) {do something}</code>. The second
one is Apple is standardizing the API for what is and is not optional.
Much of what is optional has gone from implicit to explicit, and some
is no longer optional. I've made the changes to the code here, but if
you find any more, let me know in a comment. </em>